remove dead Backtrace code
